Fixed pscid --include behaviour. (#42)
* Fixed pscid --include behaviour.

Added an ADT (CLICommand) to represent the possible command types:
NamedScriptSpecificCommand ("pscid:" script in package.json),
NamedScriptCommand (script in package.json), and
PulpCommand (eg. "pulp build").

Only the PulpCommand data constructor has an associated array of
include paths; the reasoning is described in a comment
(Options.purs:93).

These commands are transformed into the appropriate string command via
a Show instance.

* Stylistic changes.

Collapsed NamedScriptSpecificCommand and NamedScriptCommand into just
ScriptCommand, since they are treated the same everywhere.

Replaced the Show instance for CLICommand with a function (to keep with
the convention of using Show instances for debugging).

Took a couple of opportunities to use record puns.